backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
groq==1.0.0
Pillow==10.3.0
drf-spectacular==0.27.2
django-jazzmin==3.0.1
